Remove QVector from core
fixup: QVector in RenderManager

Remove QVector from AudioJack

Remove QVector from AutomatableModel

Remove QVector from ControllerConnection

Remove QVector in EffectChain

Remove QVector from EnvelopeAndLfoParameters

Remove QVector from MidiClient

Remove QVector from Mixer

Remove QVector in Note

Remove QVector in PeakController

Remove QVector from PluginFactory

Remove QVector from Track

Remove QVector from TrackContainer

Remove QVector from AutomationClip

Remove QVector from Song

Adapt non-core to QVector removal changes

Adjust panning to use std::abs
Without this, the code will not compile due to the QVector changes.

Include <vector> for MidiClient

Adapt QVector changes to PeakControllerEffect

Fix std::vector insert in Song::automatedValuesAt

Remove QVector from InstrumentFunctionNoteStacking::ChordTable

* Use std::vector and do not inherit it
* Make necessary adjustments to other parts of the codebase

Apply formatting

Use using instead of typedef where applicable

Remove QVector from ConfigManager
* Return empty QString instead of empty static QString& instead

Include <vector> where applicable

+++ src/gui/MixerView.cpp
@@ -428,8 +428,8 @@ void MixerView::deleteChannel(int index)
 void MixerView::deleteUnusedChannels()
 {
 	TrackContainer::TrackList tracks;
-	tracks += Engine::getSong()->tracks();
-	tracks += Engine::patternStore()->tracks();
+	tracks.insert(tracks.begin(), Engine::getSong()->tracks().begin(), Engine::getSong()->tracks().end());
+	tracks.insert(tracks.begin(), Engine::patternStore()->tracks().begin(), Engine::patternStore()->tracks().end());
 
 	std::vector<bool> inUse(m_mixerChannelViews.size(), false);
 
@@ -455,8 +455,7 @@ void MixerView::deleteUnusedChannels()
 	//Check all channels except master, delete those with no incoming sends
 	for(int i = m_mixerChannelViews.size()-1; i > 0; --i)
 	{
-		if (!inUse[i] && Engine::mixer()->mixerChannel(i)->m_receives.isEmpty())
-		{ deleteChannel(i); }
+		if (!inUse[i] && Engine::mixer()->mixerChannel(i)->m_receives.empty()) { deleteChannel(i); }
 	}
 }
